service-python/exportExcel/requirements.txt
Flask==0.12.1
Flask-SQLAlchemy==2.2
Mako==1.0.6
packaging==16.8
psycopg2-binary==2.8.4
psycopg2==2.7.1
pyparsing==2.2.0
python-editor==1.0.3
six==1.10.0
SQLAlchemy==1.1.9
pymongo==3.11.0
python-dotenv==0.10.3